requirements.txt
Package                 Version   
----------------------- ----------
absl-py                 0.9.0     
alembic                 1.4.2     
algoliasearch           2.2.0     
altgraph                0.17      
appnope                 0.1.0     
astor                   0.8.1     
attrs                   19.3.0    
backcall                0.1.0     
beautifulsoup4          4.8.2     
bert-for-tf2            0.12.7    
bleach                  3.1.0     
boto                    2.49.0    
boto3                   1.11.17   
botocore                1.14.17   
cachetools              4.0.0     
certifi                 2019.11.28
chardet                 3.0.4     
click                   7.1.1     
cloudpickle             1.3.0     
configparser            5.0.0     
cycler                  0.10.0    
decorator               4.4.1     
defusedxml              0.6.0     
dill                    0.3.1.1   
distro                  1.4.0     
dnspython               1.16.0    
docutils                0.15.2    
elasticsearch           7.6.0     
entrypoints             0.3       
et-xmlfile              1.0.1     
fasttext                0.9.1     
Flask                   1.1.2     
Flask-Cors              3.0.8     
Flask-Elasticsearch     0.2.5     
Flask-Login             0.5.0     
Flask-LoginManager      1.1.6     
Flask-Migrate           2.5.3     
Flask-Session           0.3.1     
Flask-SQLAlchemy        2.4.1     
future                  0.18.2    
gast                    0.2.2     
gensim                  3.8.1     
gevent                  1.4.0     
google-auth             1.10.1    
google-auth-oauthlib    0.4.1     
google-pasta            0.1.8     
greenlet                0.4.15    
grpcio                  1.26.0    
h5py                    2.10.0    
hanlp                   2.0.0a42  
hyperopt                0.2.3     
idna                    2.8       
imageio                 2.6.1     
importlib-metadata      1.4.0     
imutils                 0.5.3     
ipykernel               5.1.3     
ipython                 7.11.1    
ipython-genutils        0.2.0     
ipywidgets              7.5.1     
itsdangerous            1.1.0     
jdcal                   1.4.1     
jedi                    0.15.2    
jieba                   0.42.1    
Jinja2                  2.10.3    
jmespath                0.9.4     
joblib                  0.14.1    
jsonschema              3.2.0     
jupyter                 1.0.0     
jupyter-client          5.3.4     
jupyter-console         6.1.0     
jupyter-core            4.6.1     
kaggle                  1.5.6     
Keras                   2.3.1     
Keras-Applications      1.0.8     
Keras-Preprocessing     1.1.0     
kiwisolver              1.1.0     
lxml                    4.4.2     
macholib                1.14      
Mako                    1.1.2     
Markdown                3.1.1     
MarkupSafe              1.1.1     
MatchZoo                2.2.0     
matplotlib              3.1.2     
mistune                 0.8.4     
more-itertools          8.1.0     
mysql                   0.0.2     
mysql-connector-python  8.0.19    
mysqlclient             1.4.6     
nbconvert               5.6.1     
nbformat                5.0.3     
networkx                2.4       
nltk                    3.4.5     
notebook                6.0.2     
numexpr                 2.7.1     
numpy                   1.18.1    
oauthlib                3.1.0     
opencv-python           4.2.0.34  
openpyxl                3.0.3     
opt-einsum              3.1.0     
packaging               20.1      
pandas                  0.25.3    
pandocfilters           1.4.2     
params-flow             0.7.4     
parso                   0.5.2     
pdf2json                0.3       
pdfminer.six            20200104  
pdfminer3k              1.3.1     
pdfplumber              0.5.16    
pexpect                 4.7.0     
pickleshare             0.7.5     
Pillow                  7.0.0     
pip                     20.0.2    
pluggy                  0.13.1    
ply                     3.11      
prometheus-client       0.7.1     
prompt-toolkit          3.0.2     
protobuf                3.11.3    
ptyprocess              0.6.0     
py                      1.8.1     
py-params               0.8.2     
pyasn1                  0.4.8     
pyasn1-modules          0.2.8     
pybind11                2.5.0     
pycosat                 0.6.3     
pycryptodome            3.9.6     
pydot                   1.4.1     
Pygments                2.5.2     
PyInstaller             3.6       
PyMySQL                 0.9.3     
pyparsing               2.4.6     
PyPDF2                  1.26.0    
PyQt5                   5.14.2    
PyQt5-sip               12.7.2    
PyQt5-Tools             5.7.dev1  
pyrsistent              0.15.7    
pytesseract             0.3.3     
pytest                  5.3.5     
python-dateutil         2.8.1     
python-docx             0.8.10    
python-editor           1.0.4     
python-slugify          4.0.0     
pytorch-pretrained-bert 0.6.2     
pytz                    2019.3    
PyWavelets              1.1.1     
PyYAML                  5.3       
pyzmq                   18.1.1    
qtconsole               4.6.0     
regex                   2020.4.4  
requests                2.22.0    
requests-oauthlib       1.3.0     
rsa                     4.0       
ruamel.yaml             0.16.5    
ruamel.yaml.clib        0.2.0     
s3transfer              0.3.3     
scikit-image            0.16.2    
scikit-learn            0.22.1    
scipy                   1.4.1     
seaborn                 0.10.0    
selenium                3.141.0   
Send2Trash              1.5.0     
sentencepiece           0.1.85    
setuptools              41.2.0    
six                     1.13.0    
smart-open              1.9.0     
sortedcontainers        2.1.0     
soupsieve               1.9.5     
SQLAlchemy              1.3.16    
tables                  3.6.1     
tabula-py               2.0.4     
tensorboard             2.1.1     
tensorflow              2.1.0     
tensorflow-estimator    2.1.0     
termcolor               1.1.0     
terminado               0.8.3     
testpath                0.4.4     
text-unidecode          1.3       
torch                   1.5.0     
tornado                 6.0.3     
tqdm                    4.41.1    
traitlets               4.3.3     
unicodecsv              0.14.1    
urllib3                 1.24.3    
Wand                    0.5.9     
wcwidth                 0.1.8     
webencodings            0.5.1     
Werkzeug                0.16.0    
wheel                   0.33.6    
widgetsnbextension      3.5.1     
wrapt                   1.11.2    
xlrd                    1.2.0     
zipp                    1.0.0     
